+++ include/openpose/producer/datumProducer.hpp
@@ -11,19 +11,19 @@
 
 namespace op
 {
-    template<typename TDatumsNoPtr>
+    template<typename TDatum>
     class DatumProducer
     {
     public:
-        explicit DatumProducer(const std::shared_ptr<Producer>& producerSharedPtr,
-                               const unsigned long long frameFirst = 0, const unsigned long long frameStep = 1,
-                               const unsigned long long frameLast = std::numeric_limits<unsigned long long>::max(),
-                               const std::shared_ptr<std::pair<std::atomic<bool>,
-                                                     std::atomic<int>>>& videoSeekSharedPtr = nullptr);
+        explicit DatumProducer(
+            const std::shared_ptr<Producer>& producerSharedPtr,
+            const unsigned long long frameFirst = 0, const unsigned long long frameStep = 1,
+            const unsigned long long frameLast = std::numeric_limits<unsigned long long>::max(),
+            const std::shared_ptr<std::pair<std::atomic<bool>, std::atomic<int>>>& videoSeekSharedPtr = nullptr);
 
         virtual ~DatumProducer();
 
-        std::pair<bool, std::shared_ptr<TDatumsNoPtr>> checkIfRunningAndGetDatum();
+        std::pair<bool, std::shared_ptr<std::vector<std::shared_ptr<TDatum>>>> checkIfRunningAndGetDatum();
 
     private:
         const unsigned long long mNumberFramesToProcess;
@@ -33,8 +33,8 @@ namespace op
         unsigned int mNumberConsecutiveEmptyFrames;
         std::shared_ptr<std::pair<std::atomic<bool>, std::atomic<int>>> spVideoSeek;
 
-        void checkIfTooManyConsecutiveEmptyFrames(unsigned int& numberConsecutiveEmptyFrames,
-                                                  const bool emptyFrame) const;
+        void checkIfTooManyConsecutiveEmptyFrames(
+            unsigned int& numberConsecutiveEmptyFrames, const bool emptyFrame) const;
 
         DELETE_COPY(DatumProducer);
     };
@@ -49,12 +49,12 @@ namespace op
 #include <openpose/producer/datumProducer.hpp>
 namespace op
 {
-    template<typename TDatumsNoPtr>
-    DatumProducer<TDatumsNoPtr>::DatumProducer(const std::shared_ptr<Producer>& producerSharedPtr,
-                                               const unsigned long long frameFirst, const unsigned long long frameStep,
-                                               const unsigned long long frameLast,
-                                               const std::shared_ptr<std::pair<std::atomic<bool>,
-                                                                               std::atomic<int>>>& videoSeekSharedPtr) :
+    template<typename TDatum>
+    DatumProducer<TDatum>::DatumProducer(
+        const std::shared_ptr<Producer>& producerSharedPtr,
+        const unsigned long long frameFirst, const unsigned long long frameStep,
+        const unsigned long long frameLast,
+        const std::shared_ptr<std::pair<std::atomic<bool>, std::atomic<int>>>& videoSeekSharedPtr) :
         mNumberFramesToProcess{(frameLast != std::numeric_limits<unsigned long long>::max()
                                 ? frameLast - frameFirst : frameLast)},
         spProducer{producerSharedPtr},
@@ -92,17 +92,17 @@ namespace op
         }
     }
 
-    template<typename TDatumsNoPtr>
-    DatumProducer<TDatumsNoPtr>::~DatumProducer()
+    template<typename TDatum>
+    DatumProducer<TDatum>::~DatumProducer()
     {
     }
 
-    template<typename TDatumsNoPtr>
-    std::pair<bool, std::shared_ptr<TDatumsNoPtr>> DatumProducer<TDatumsNoPtr>::checkIfRunningAndGetDatum()
+    template<typename TDatum>
+    std::pair<bool, std::shared_ptr<std::vector<std::shared_ptr<TDatum>>>> DatumProducer<TDatum>::checkIfRunningAndGetDatum()
     {
         try
         {
-            auto datums = std::make_shared<TDatumsNoPtr>();
+            auto datums = std::make_shared<std::vector<std::shared_ptr<TDatum>>>();
             // Check last desired frame has not been reached
             if (mNumberFramesToProcess != std::numeric_limits<unsigned long long>::max()
                 && mGlobalCounter > mNumberFramesToProcess)
@@ -137,52 +137,54 @@ namespace op
                 {
                     datums->resize(cvMats.size());
                     // Datum cannot be assigned before resize()
-                    auto& datum = (*datums)[0];
+                    auto& datumPtr = (*datums)[0];
+                    datumPtr = std::make_shared<TDatum>();
                     // Filling first element
-                    std::swap(datum.name, nextFrameName);
-                    datum.frameNumber = nextFrameNumber;
-                    datum.cvInputData = cvMats[0];
+                    std::swap(datumPtr->name, nextFrameName);
+                    datumPtr->frameNumber = nextFrameNumber;
+                    datumPtr->cvInputData = cvMats[0];
                     if (!cameraMatrices.empty())
                     {
-                        datum.cameraMatrix = cameraMatrices[0];
-                        datum.cameraExtrinsics = cameraExtrinsics[0];
-                        datum.cameraIntrinsics = cameraIntrinsics[0];
+                        datumPtr->cameraMatrix = cameraMatrices[0];
+                        datumPtr->cameraExtrinsics = cameraExtrinsics[0];
+                        datumPtr->cameraIntrinsics = cameraIntrinsics[0];
                     }
                     // Image integrity
-                    if (datum.cvInputData.channels() != 3)
+                    if (datumPtr->cvInputData.channels() != 3)
                     {
                         const std::string commonMessage{"Input images must be 3-channel BGR."};
                         // Grey to RGB if required
-                        if (datum.cvInputData.channels() == 1)
+                        if (datumPtr->cvInputData.channels() == 1)
                         {
                             log(commonMessage + " Converting grey image into BGR.", Priority::High);
-                            cv::cvtColor(datum.cvInputData, datum.cvInputData, CV_GRAY2BGR);
+                            cv::cvtColor(datumPtr->cvInputData, datumPtr->cvInputData, CV_GRAY2BGR);
                         }
                         else
                             error(commonMessage, __LINE__, __FUNCTION__, __FILE__);
                     }
-                    datum.cvOutputData = datum.cvInputData;
+                    datumPtr->cvOutputData = datumPtr->cvInputData;
                     // Resize if it's stereo-system
                     if (datums->size() > 1)
                     {
                         // Stereo-system: Assign all cv::Mat
                         for (auto i = 1u ; i < datums->size() ; i++)
                         {
-                            auto& datumI = (*datums)[i];
-                            datumI.name = datum.name;
-                            datumI.frameNumber = datum.frameNumber;
-                            datumI.cvInputData = cvMats[i];
-                            datumI.cvOutputData = datumI.cvInputData;
+                            auto& datumIPtr = (*datums)[i];
+                            datumIPtr = std::make_shared<TDatum>();
+                            datumIPtr->name = datumPtr->name;
+                            datumIPtr->frameNumber = datumPtr->frameNumber;
+                            datumIPtr->cvInputData = cvMats[i];
+                            datumIPtr->cvOutputData = datumIPtr->cvInputData;
                             if (cameraMatrices.size() > i)
                             {
-                                datumI.cameraMatrix = cameraMatrices[i];
-                                datumI.cameraExtrinsics = cameraExtrinsics[i];
-                                datumI.cameraIntrinsics = cameraIntrinsics[i];
+                                datumIPtr->cameraMatrix = cameraMatrices[i];
+                                datumIPtr->cameraExtrinsics = cameraExtrinsics[i];
+                                datumIPtr->cameraIntrinsics = cameraIntrinsics[i];
                             }
                         }
                     }
                     // Check producer is running
-                    if (!datumProducerRunning || (*datums)[0].cvInputData.empty())
+                    if (!datumProducerRunning || (*datums)[0]->cvInputData.empty())
                         datums = nullptr;
                     // Increase counter if successful image
                     if (datums != nullptr)
@@ -195,13 +197,13 @@ namespace op
         catch (const std::exception& e)
         {
             error(e.what(), __LINE__, __FUNCTION__, __FILE__);
-            return std::make_pair(false, std::make_shared<TDatumsNoPtr>());
+            return std::make_pair(false, std::make_shared<std::vector<std::shared_ptr<TDatum>>>());
         }
     }
 
-    template<typename TDatumsNoPtr>
-    void DatumProducer<TDatumsNoPtr>::checkIfTooManyConsecutiveEmptyFrames(unsigned int& numberConsecutiveEmptyFrames,
-                                                                           const bool emptyFrame) const
+    template<typename TDatum>
+    void DatumProducer<TDatum>::checkIfTooManyConsecutiveEmptyFrames(
+        unsigned int& numberConsecutiveEmptyFrames, const bool emptyFrame) const
     {
         numberConsecutiveEmptyFrames = (emptyFrame ? numberConsecutiveEmptyFrames+1 : 0);
         const auto threshold = 500u;
@@ -210,7 +212,7 @@ namespace op
                   __LINE__, __FUNCTION__, __FILE__);
     }
 
-    extern template class DatumProducer<DATUM_BASE_NO_PTR>;
+    extern template class DatumProducer<BASE_DATUM>;
 }